adds authentication verifying to controller methods and fixes errors

+++ API/ChecklistAPI/src/main/java/pt/isel/daw/g5/ChecklistAPI/controller/ChecklistTemplateController.java
@@ -47,7 +47,7 @@ public class ChecklistTemplateController {
     @PostMapping
     public String postChecklistTemplate(@RequestBody ChecklistTemplate checklistTemplate) {
         Optional<User> rui = userRepository.findById("Rui");
-        checklistTemplate.setUserName(rui.get()); //TODO verificar autenticação
+        checklistTemplate.setUser(rui.get()); //TODO verificar autenticação
         checklistTemplateRepository.save(checklistTemplate);
         return "Ok";
     }
@@ -64,7 +64,7 @@ public class ChecklistTemplateController {
 
         ChecklistTemplate checklistTemplate = optionalChecklistTemplate.get();
 
-        if (!checklistTemplate.getUserName().getUsername().equals(username)){
+        if (!checklistTemplate.getUser().getUsername().equals(username)){
             InvalidParams notIncludedUser = new InvalidParams("username", "username is invalid");
             ProblemJSON problemJSON = new ProblemJSON("/authentication-error", "Invalid User.", 403, "The user provided does not have access to this list", request.getRequestURI(), new InvalidParams[]{notIncludedUser});
             throw new NotAuthenticatedException(problemJSON);
@@ -83,7 +83,7 @@ public class ChecklistTemplateController {
         if (!optionalChecklistTemplate.isPresent())
             throw new NotFoundException();
 
-        if (!optionalChecklistTemplate.get().getUserName().getUsername().equals(username)){
+        if (!optionalChecklistTemplate.get().getUser().getUsername().equals(username)){
             InvalidParams notIncludedUser = new InvalidParams("username", "username is invalid");
             ProblemJSON problemJSON = new ProblemJSON("/authentication-error", "Invalid User.", 403, "The user provided does not have access to this list", request.getRequestURI(), new InvalidParams[]{notIncludedUser});
             throw new NotAuthenticatedException(problemJSON);
@@ -104,13 +104,13 @@ public class ChecklistTemplateController {
         if (!optionalChecklistTemplate.isPresent())
             throw new NotFoundException();
 
-        if (!optionalChecklistTemplate.get().getUserName().getUsername().equals(username)){
+        if (!optionalChecklistTemplate.get().getUser().getUsername().equals(username)){
             InvalidParams notIncludedUser = new InvalidParams("username", "username is invalid");
             ProblemJSON problemJSON = new ProblemJSON("/authentication-error", "Invalid User.", 403, "The user provided does not have access to this list", request.getRequestURI(), new InvalidParams[]{notIncludedUser});
             throw new NotAuthenticatedException(problemJSON);
         }
 
-        checklistTemplate.setUserName(userRepository.findById(username).get());
+        checklistTemplate.setUser(userRepository.findById(username).get());
         checklistTemplate.setId(checklisttemplate_id);
         checklistTemplateRepository.save(checklistTemplate);
         return "OK";
@@ -118,9 +118,22 @@ public class ChecklistTemplateController {
 
     @GetMapping(path = "/{checklisttemplate_id}/templateitems", produces = "application/vnd.collection+json")
     public OutTemplateItems getTemplateItems(@PathVariable("checklisttemplate_id") int checklistTemplateId,
-                                             @RequestParam(value = "page", defaultValue = "0") int page){
+                                             @RequestParam(value = "page", defaultValue = "0") int page,
+                                             HttpServletRequest request){
 
-        if(!checklistTemplateRepository.existsById(checklistTemplateId)) { /* TODO ERROR HANDLING*/ }
+        if(!checklistTemplateRepository.existsById(checklistTemplateId))
+            throw new NotFoundException();
+
+        String username = (String) request.getAttribute("Username");
+        Optional<ChecklistTemplate> optionalChecklistTemplate = checklistTemplateRepository.findById(checklistTemplateId);
+        ChecklistTemplate template = optionalChecklistTemplate.get();
+
+        if(!template.getUser().getUsername().equals(username)){
+            InvalidParams notIncludedUser = new InvalidParams("username", "username is invalid");
+            ProblemJSON problemJSON = new ProblemJSON("/authentication-error", "Invalid User.", 403, "The user provided does not have access to this list", request.getRequestURI(), new InvalidParams[]{notIncludedUser});
+            throw new NotAuthenticatedException(problemJSON);
+        }
+
         Page<TemplateItem> itemPage = templateItemRepository.findAllByChecklistTemplate_Id(checklistTemplateId, PageRequest.of(page, PAGE_SIZE));
         return new OutTemplateItems(itemPage, checklistTemplateId);
     }
@@ -136,7 +149,7 @@ public class ChecklistTemplateController {
         if (!optionalChecklistTemplate.isPresent())
             throw new NotFoundException();
 
-        if (!optionalChecklistTemplate.get().getUserName().getUsername().equals(username)){
+        if (!optionalChecklistTemplate.get().getUser().getUsername().equals(username)){
             InvalidParams notIncludedUser = new InvalidParams("username", "username is invalid");
             ProblemJSON problemJSON = new ProblemJSON("/authentication-error", "Invalid User.", 403, "The user provided does not have access to this list", request.getRequestURI(), new InvalidParams[]{notIncludedUser});
             throw new NotAuthenticatedException(problemJSON);
@@ -151,7 +164,21 @@ public class ChecklistTemplateController {
 
     @GetMapping(path = "/{checklisttemplate_id}/templateitems/{templateitem_id}", produces = "application/vnd.siren+json")
     public OutTemplateItem getTemplateItem(@PathVariable("checklisttemplate_id") int checklistTemplateId,
-                                           @PathVariable("templateitem_id") int templateItemId){
+                                           @PathVariable("templateitem_id") int templateItemId,
+                                           HttpServletRequest request){
+
+        if(!checklistTemplateRepository.existsById(checklistTemplateId) || !templateItemRepository.existsById(templateItemId))
+            throw new NotFoundException();
+
+        String username = (String) request.getAttribute("Username");
+        Optional<ChecklistTemplate> optionalChecklistTemplate = checklistTemplateRepository.findById(checklistTemplateId);
+        ChecklistTemplate template = optionalChecklistTemplate.get();
+
+        if(!template.getUser().getUsername().equals(username)){
+            InvalidParams notIncludedUser = new InvalidParams("username", "username is invalid");
+            ProblemJSON problemJSON = new ProblemJSON("/authentication-error", "Invalid User.", 403, "The user provided does not have access to this list", request.getRequestURI(), new InvalidParams[]{notIncludedUser});
+            throw new NotAuthenticatedException(problemJSON);
+        }
 
         TemplateItem templateItem = templateItemRepository.findById(templateItemId).get();
         DatabaseTemplateItem databaseTemplateItem = new DatabaseTemplateItem(templateItem);
@@ -161,20 +188,47 @@ public class ChecklistTemplateController {
     @PutMapping("/{checklisttemplate_id}/templateitems/{templateitem_id}")
     public String updateTemplateItem(@PathVariable("checklisttemplate_id") int checklistTemplateId,
                                      @PathVariable("templateitem_id") int templateItemId,
-                                     @RequestBody TemplateItem templateItem){
+                                     @RequestBody DatabaseTemplateItem templateItem,
+                                     HttpServletRequest request){
 
         if(!checklistTemplateRepository.existsById(checklistTemplateId) || !templateItemRepository.existsById(templateItemId))
-            return "ERROR"; // TODO ???
-        templateItemRepository.save(templateItem);
+            throw new NotFoundException();
+
+        String username = (String) request.getAttribute("Username");
+        Optional<ChecklistTemplate> optionalChecklistTemplate = checklistTemplateRepository.findById(checklistTemplateId);
+        ChecklistTemplate template = optionalChecklistTemplate.get();
+
+        if(!template.getUser().getUsername().equals(username)){
+            InvalidParams notIncludedUser = new InvalidParams("username", "username is invalid");
+            ProblemJSON problemJSON = new ProblemJSON("/authentication-error", "Invalid User.", 403, "The user provided does not have access to this list", request.getRequestURI(), new InvalidParams[]{notIncludedUser});
+            throw new NotAuthenticatedException(problemJSON);
+        }
+
+        TemplateItem updatedItem = templateItemRepository.findById(templateItemId).get();
+        updatedItem.setName(templateItem.getName());
+        updatedItem.setDescription(templateItem.getDescription());
+        templateItemRepository.save(updatedItem);
         return "OK";
     }
 
     @DeleteMapping("/{checklisttemplate_id}/templateitems/{templateitem_id}")
     public String deleteTemplateItem(@PathVariable("checklisttemplate_id") int checklistTemplateId,
-                                     @PathVariable("templateitem_id") int templateItemId){
+                                     @PathVariable("templateitem_id") int templateItemId,
+                                     HttpServletRequest request){
 
         if(!checklistTemplateRepository.existsById(checklistTemplateId))
-            return "ERROR"; // TODO ???
+            throw new NotFoundException();
+
+        String username = (String) request.getAttribute("Username");
+        Optional<ChecklistTemplate> optionalChecklistTemplate = checklistTemplateRepository.findById(checklistTemplateId);
+        ChecklistTemplate template = optionalChecklistTemplate.get();
+
+        if(!template.getUser().getUsername().equals(username)){
+            InvalidParams notIncludedUser = new InvalidParams("username", "username is invalid");
+            ProblemJSON problemJSON = new ProblemJSON("/authentication-error", "Invalid User.", 403, "The user provided does not have access to this list", request.getRequestURI(), new InvalidParams[]{notIncludedUser});
+            throw new NotAuthenticatedException(problemJSON);
+        }
+
         templateItemRepository.deleteById(templateItemId);
         return "OK";
     }
